refactor: move `exit_fullscreen` to `fullscreen` subfolder
