+++ components/script/dom/fullscreen/fullscreen.rs
@@ -8,8 +8,9 @@ use crate::dom::bindings::codegen::Bindings::WindowBinding::WindowMethods;
 use crate::dom::bindings::error::Error;
 use crate::dom::bindings::inheritance::Castable;
 use crate::dom::bindings::refcounted::{Trusted, TrustedPromise};
+use crate::dom::bindings::reflector::DomGlobal;
 use crate::dom::document::document::Document;
-use crate::dom::element::{Element, ElementPerformFullscreenEnter};
+use crate::dom::element::{Element, ElementPerformFullscreenEnter, ElementPerformFullscreenExit};
 use crate::dom::node::NodeTraits;
 use crate::dom::promise::Promise;
 use crate::dom::types::HTMLDialogElement;
@@ -132,6 +133,56 @@ impl Document {
         );
         let msg = MainThreadScriptMsg::Common(script_msg);
         self.window().main_thread_script_chan().send(msg).unwrap();
+
+        promise
+    }
+
+    /// <https://fullscreen.spec.whatwg.org/#exit-fullscreen>
+    pub(crate) fn exit_fullscreen(&self, can_gc: CanGc) -> Rc<Promise> {
+        let global = self.global();
+
+        // Step 1
+        // > Let promise be a new promise
+        let in_realm_proof = AlreadyInRealm::assert::<crate::DomTypeHolder>();
+        let promise = Promise::new_in_current_realm(InRealm::Already(&in_realm_proof), can_gc);
+
+        // Step 2
+        // > If doc is not fully active or doc’s fullscreen element is null, then reject promise with a TypeError exception and return promise.
+        if !self.is_fully_active() || self.fullscreen_element().is_none() {
+            promise.reject_error(
+                Error::Type(
+                    c"No fullscreen element to exit or document is not fully active".to_owned(),
+                ),
+                can_gc,
+            );
+            return promise;
+        }
+
+        // TODO(#42067): Implement step 3-7, handling fullscreen's propagation across navigables.
+
+        let element = self.fullscreen_element().unwrap();
+        let window = self.window();
+
+        // Step 10
+        // > If resize is true, resize doc’s viewport to its "normal" dimensions.
+        // TODO(#21600): Improve spec compliance of steps 8-15 paralelism.
+        let event = EmbedderMsg::NotifyFullscreenStateChanged(self.webview_id(), false);
+        self.send_to_embedder(event);
+
+        // Step 8
+        // > Return promise, and run the remaining steps in parallel.
+        let trusted_element = Trusted::new(&*element);
+        let trusted_promise = TrustedPromise::new(promise.clone());
+        let handler = ElementPerformFullscreenExit::new(trusted_element, trusted_promise);
+        let pipeline_id = Some(global.pipeline_id());
+        let script_msg = CommonScriptMsg::Task(
+            ScriptThreadEventCategory::ExitFullscreen,
+            handler,
+            pipeline_id,
+            TaskSourceName::DOMManipulation,
+        );
+        let msg = MainThreadScriptMsg::Common(script_msg);
+        window.main_thread_script_chan().send(msg).unwrap();
 
         promise
     }
