+++ components/script/dom/fullscreen/mod.rs
@@ -2,4 +2,5 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
+#[allow(clippy::module_inception, reason = "The interface name is fullscreen")]
 pub mod fullscreen;
